fix double <<

+++ srcs/exec/handle_heredoc.c
@@ -6,7 +6,7 @@
 /*   By: ijaber <ijaber@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/04 13:33:10 by ijaber            #+#    #+#             */
-/*   Updated: 2024/10/16 16:20:36 by ijaber           ###   ########.fr       */
+/*   Updated: 2024/10/28 14:23:24 by ijaber           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,23 +26,12 @@ char	*get_temp_filename(t_data *data)
 	return (filename);
 }
 
-int	create_heredoc(char *delimiter, t_data *data)
+static int	read_heredoc_content(int fd, const char *real_delimiter,
+		char *delimiter)
 {
-	static size_t	count_line;
-	const char		*real_delimiter = ft_strjoin(delimiter, "\n");
-	char			*filename;
-	int				fd;
+	static size_t	count_line = 1;
 	char			*line;
 
-	count_line = 1;
-	filename = get_temp_filename(data);
-	if (!real_delimiter)
-		handle_malloc_error("heredoc", data);
-	if (!filename)
-		return (-1);
-	fd = ft_open_outfile(filename, O_APPEND, data);
-	if (fd == -1)
-		return (-1);
 	while (1)
 	{
 		setup_heredoc_signals();
@@ -56,38 +45,34 @@ int	create_heredoc(char *delimiter, t_data *data)
 %d delimited by end-of-file (wanted `%s')\n",
 							count_line,
 							delimiter);
-			break ;
+			return (0);
 		}
 		write(fd, line, strlen(line));
 		count_line++;
 	}
+	return (0);
+}
+
+int	create_heredoc(char *delimiter, t_data *data)
+{
+	const char	*real_delimiter;
+	char		*filename;
+	int			fd;
+
+	real_delimiter = ft_strjoin(delimiter, "\n");
+	filename = get_temp_filename(data);
+	if (!real_delimiter)
+		handle_malloc_error("heredoc", data);
+	if (!filename)
+		return (-1);
+	fd = ft_open_outfile(filename, O_APPEND, data);
+	if (fd == -1)
+		return (-1);
+	read_heredoc_content(fd, real_delimiter, delimiter);
 	ft_close(fd);
 	fd = open(filename, O_RDONLY);
 	unlink(filename);
 	return (fd);
-}
-
-int	handle_heredoc_in(t_ast_node *node, t_data *data)
-{
-	int	here_doc_fd;
-
-	here_doc_fd = create_heredoc(node->right->args[0], data);
-	if (here_doc_fd == -1)
-		free_and_exit(-1);
-	data->backup_stdin = dup(STDIN_FILENO);
-	if (dup2(here_doc_fd, STDIN_FILENO) == -1)
-	{
-		ft_fprintf(2, "minishell: Error when trying to dup2\n");
-		(ft_close(here_doc_fd), ft_close(data->backup_stdin));
-		if (data->is_child == 1)
-			free_and_exit(-1);
-		return (0);
-	}
-	ft_close(here_doc_fd);
-	execute_ast(node->left, data);
-	dup2(data->backup_stdin, STDIN_FILENO);
-	ft_close(data->backup_stdin);
-	return (0);
 }
 
 int	handle_heredoc_out(t_ast_node *node, t_data *data)
